ua-tool: Pass the token via a private temporary attach config file
We're currently calling the pro client with attach parameter in plain
text, but this leads to leaking such private information to any user in
the system (via simple ps).

So use a temporary file instead. This is safe because it's created by
default with -rw------- only permissions, so it won't be readable by
anyone else.

We could use an extra layer of protection saving it in the runtime
directory, but this is already enough.

LP: #2068944